[DD-1392] separated JSON schema parser from XML schema parser
